requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
PyJWT==2.3.0
